simulation/requirements.txt
databricks-sdk>=0.45.0
faker>=24.0.0
pandas>=2.0.0
pyarrow>=14.0.0
